Change u64 to usize.

+++ crates/blockifier/src/fee/eth_gas_constants.rs
@@ -1,5 +1,5 @@
-pub const WORD_WIDTH: u64 = 32;
-pub const GAS_PER_LOG: u64 = 375;
-pub const GAS_PER_LOG_DATA_BYTE: u64 = 8;
-pub const GAS_PER_LOG_DATA_WORD: u64 = GAS_PER_LOG_DATA_BYTE * WORD_WIDTH;
-pub const GAS_PER_LOG_TOPIC: u64 = 375;
+pub const WORD_WIDTH: usize = 32;
+pub const GAS_PER_LOG: usize = 375;
+pub const GAS_PER_LOG_DATA_BYTE: usize = 8;
+pub const GAS_PER_LOG_DATA_WORD: usize = GAS_PER_LOG_DATA_BYTE * WORD_WIDTH;
+pub const GAS_PER_LOG_TOPIC: usize = 375;
